First Commit
What exists:
Able to add "TestTypes" using macros to a global link-list of "TestTypes" which can be accessed using the temporary methods provided.

NOTE: Have to clean up ALOT of stuff and figure out how to store the TestTypes in memory so it is more efficient to flush out. How to re-make the data structure to have it support registering each test run and have it show run "hierarchies".